requirements.txt
Pillow
jupyter
matplotlib
pandas
regex
scikit-image
scikit-learn
scipy
spacy
tqdm


#torch~=1.9.0+cu111
#torchvision~=0.10.0+cu111
